chore(pollux): split project in two crates
