+++ Cargo.toml
@@ -1,9 +1,12 @@
 [workspace]
-members = ["crates/*"]
+members = [
+    "crates/pollux-cli",
+    "crates/pollux-core",
+    "crates/xtasks",
+]
 resolver = "2"
 
 [workspace.package]
-version = "0.1.0"
 edition = "2024"
 rust-version = "1.90.0"
 authors = ["dotanuki.labs@gmail.com"]
@@ -14,10 +17,10 @@ documentation = "https://github.com/dotanuki-labs/pollux"
 readme = "README.md"
 license = "MIT"
 categories = ["command-line-utilities", "security"]
+version = "0.1.0"
 
 [workspace.dependencies]
-pollux = { path = "crates/pollux" }
-xtasks = { path = "crates/xtasks" }
+pollux-core = { path = "crates/pollux-core", version = "0.1.0" }
 
 anyhow = "=1.0.100"
 assertor = "=0.0.4"
