+++ crates/pollux-cli/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "pollux"
+description = "Get provenance and build reproducibility insights for your Rust supply chain"
+authors.workspace = true
+edition.workspace = true
+documentation.workspace = true
+homepage.workspace = true
+license.workspace = true
+readme.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+publish = true
+
+[dependencies]
+pollux-core = { path = "../pollux-core", version = "0.1.0" }
+anyhow.workspace = true
+tikv-jemallocator.workspace = true
+tokio.workspace = true
+
+[dev-dependencies]
+assertor.workspace = true
+assert_cmd.workspace = true
+predicates.workspace = true
+temp-dir.workspace = true
