+++ crates/pollux-cli/src/main.rs
@@ -1,12 +1,8 @@
 // Copyright 2025 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
-mod core;
-mod infra;
-mod ioc;
-mod pollux;
-
-use crate::infra::cli;
+use pollux_core::factory;
+use pollux_core::infra::cli;
 use tikv_jemallocator::Jemalloc;
 
 #[global_allocator]
@@ -17,7 +13,7 @@ async fn main() -> anyhow::Result<()> {
     cli::troubleshooting::setup_troubleshooting();
     let (task, turnoff_colors) = cli::parsing::parse_arguments()?;
 
-    let pollux = ioc::create_pollux(turnoff_colors);
+    let pollux = factory::create_pollux(turnoff_colors);
     pollux.execute(task).await?;
 
     Ok(())
