+++ crates/pollux-core/src/domain.rs
@@ -1,0 +1,6 @@
+// Copyright 2025 Dotanuki Labs
+// SPDX-License-Identifier: MIT
+
+pub(crate) mod analysers;
+pub(crate) mod interfaces;
+pub(crate) mod models;
