+++ crates/pollux-core/src/domain/analysers.rs
