+++ crates/pollux-core/src/domain/models.rs
