+++ crates/pollux-core/src/infra.rs
@@ -1,6 +1,7 @@
 // Copyright 2025 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
-pub mod analysers;
-pub mod interfaces;
-pub mod models;
+pub mod cli;
+
+pub(crate) mod caching;
+pub(crate) mod networking;
