+++ crates/pollux-core/src/infra/cli/troubleshooting.rs
