+++ crates/pollux-core/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright 2025 Dotanuki Labs
 // SPDX-License-Identifier: MIT
 
-pub mod caching;
-pub mod cli;
-pub mod networking;
+pub mod domain;
+pub mod factory;
+pub mod infra;
+pub mod pollux;
