+++ crates/xtasks/Cargo.toml
@@ -10,6 +10,7 @@ readme.workspace = true
 repository.workspace = true
 rust-version.workspace = true
 version.workspace = true
+publish = false
 
 [dependencies]
 anyhow.workspace = true
